Refactor ValidatorModal
